Bugfix in seismic anomalies calculation (#17)
# Description
- Corrected number of atom in formula unit for calculation of `_E_th_dT`
for Bm and CaPv
- Changed from `k_bm_0` to `k_v_bm` in calculation of alpha for Bm
- Corrected calculation of `rho_capv`
- Corrected processing of files in `_calc_seismic_anomalies` to ignore
already processed files
- Corrected dummy value for `rho_fp` when no Fp is present, otherwise it
results to NaN value
- Changed from `feo` to `ratio_fe` in calculation of composition of
Bridgmanite in term of components

# Note
The current calculation of the seismic properties were verified against
the matlab version used in Vilella et al. (2021) so it should be
correct.
They are however still issues with the calculation of the spin
configuration and mineral composition.